+++ src/app/app.component.html
@@ -2,7 +2,7 @@
   <ion-menu side="start" menuId="first" contentId="content1">
     <ion-header>
       <ion-toolbar>
-        <ion-title>Tech. Alloc. Menu</ion-title>
+        <ion-title>Menu</ion-title>
       </ion-toolbar>
     </ion-header>
     <ion-content>
@@ -15,6 +15,9 @@
       </ion-menu-toggle>
       </ion-list>
     </ion-content>
+    <ion-footer>
+        Developed by:'dimensionaltechs@gmail.com'
+    </ion-footer>
   </ion-menu>
 <ion-router-outlet id="content1"></ion-router-outlet>
 </ion-app>
