+++ src/app/home/home.component.html
@@ -1,0 +1,31 @@
+<ion-app>
+    <ion-header>
+        <ion-toolbar>
+            <ion-buttons slot="start">
+                <ion-menu-button autoHide="false"></ion-menu-button>
+            </ion-buttons>
+            <ion-title>Eskom Tech. Allocation</ion-title>
+        </ion-toolbar>
+    </ion-header>
+
+    <ion-content class="ion-padding">
+        <form [formGroup]="simple_form">
+        <!--form-->
+            <ion-item>
+                <ion-label>Latitude: </ion-label>
+                <ion-input type="text" formControlName="latitude" value={{lati}}></ion-input>
+            </ion-item>
+            <ion-item>
+                <ion-label>Longitude: </ion-label>
+                <ion-input type="text" formControlName="longitude" value={{long}}></ion-input>
+            </ion-item>
+            <ion-item>
+                <ion-label position="floating">Please enter status of you task(s)</ion-label>
+                <ion-textarea rows="5" formControlName="tech_message"></ion-textarea>
+            </ion-item>
+            <ion-button class="submit-btn" expand="block" (click)="addToDB(simple_form.value)" type="submit">Update My Location</ion-button>
+        </form>
+        <!-- Google Maps-->
+        <div #mapElement class="map"></div>
+        </ion-content>
+</ion-app>
