+++ src/app/home/home.component.scss
@@ -1,0 +1,37 @@
+.submit-btn {
+    margin-top: 40px;
+}
+
+ion-input {
+    // --background: #3880ff;
+    --background: #fff;
+    --color: black;
+}
+
+.amb_inputs {
+    width: 100%;
+    padding: 12px;
+    color: black;
+    border-radius: 4px;
+    margin-top: 4px;
+    resize: none;
+    border-bottom: 1px solid black;
+}
+
+#map_canvas {
+    width: 100%;
+    height: 100%;
+}
+
+agm-map {
+    align-content: center;
+    width: 50%;
+    height: 300px;
+}
+//Delete agm-map if this below works fine
+.map {
+    //    height: 100%;
+    align-content: center;
+    width: 100%;
+    height: 100%;
+}
