--- src/app/home/home.page.html
@@ -1,34 +1,0 @@
-<ion-app>
-  <ion-header>
-    <ion-toolbar>
-      <ion-buttons slot="start">
-        <ion-menu-button autoHide="false"></ion-menu-button>
-      </ion-buttons>
-      <ion-title>Abulance Allocation</ion-title>
-    </ion-toolbar>
-  </ion-header>
-  
-  <ion-content class="ion-padding">
-    <form [formGroup]="simple_form">
-      <ion-item>
-        <ion-label>Latitude: </ion-label>
-        <ion-input type="text" formControlName="latitude" value={{lati}}></ion-input>
-      </ion-item>
-      <ion-item>
-        <ion-label>Longitude: </ion-label>
-        <ion-input type="text" formControlName="longitude" value={{long}}></ion-input>
-       </ion-item>
-      <ion-item>
-        <ion-label position="floating">Please enter status of you task(s)</ion-label>
-        <ion-textarea rows="5" formControlName="tech_message"></ion-textarea>
-      </ion-item>
-      <ion-button class="submit-btn" expand="block" (click)="addToDB(simple_form.value)" type="submit">Update My Location</ion-button>
-    </form>
-    <!-- Map container -->
-    <div id="map_canvas">
-      <!-- <agm-map [latitude]="lati" [longitude]="long">
-        <agm-marker [latitude]="lati" [longitude]="long"></agm-marker>
-      </agm-map> -->
-    </div>
-  </ion-content>
-</ion-app>
