+++ src/index.html
@@ -16,6 +16,7 @@
   <!-- add to homescreen for ios -->
   <meta name="apple-mobile-web-app-capable" content="yes" />
   <meta name="apple-mobile-web-app-status-bar-style" content="black" />
+  <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBaZWRrDJq_NdI6MZMkQ4JDm96M7ckIjDE"></script>
 </head>
 
 <body>
